rts: don't export atomic_inc64
This makes it easier for packges to use GHC 9.6.7 because the
symbols exported by the rts are the same as in 9.6.6.

This means that we define the symbol in the ghc library
instead, where we have to pessimistically assume a threaded
rts.

(cherry picked from commit 6205e04ae920776f282df9a428829d9e90a9c776)
Signed-off-by: Han Gao <rabenda.cn@gmail.com>